# Document/swagger-insure.v1.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "RFQ"
    },
    "servers": [
        {
            "url": "http://192.168.3.56:9763",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/templates/update/{id}": {
            "put": {
                "tags": [
                    "template-controller"
                ],
                "operationId": "updateTemplate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "templateName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "templateType",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permissions",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "templatefile": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemplateDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/refreshToken": {
            "post": {
                "tags": [
                    "user-login-controller"
                ],
                "operationId": "getRefreshedToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshAuthDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/authenticate": {
            "post": {
                "tags": [
                    "user-login-controller"
                ],
                "operationId": "getValidToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticationDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/templates/create": {
            "post": {
                "tags": [
                    "template-controller"
                ],
                "operationId": "createTemplate",
                "parameters": [
                    {
                        "name": "templateName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "templateType",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permissions",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "productId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "templateFile"
                                ],
                                "type": "object",
                                "properties": {
                                    "templateFile": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/templates/{id}": {
            "get": {
                "tags": [
                    "template-controller"
                ],
                "operationId": "getFileDataById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/templates/files/byType/{type}": {
            "get": {
                "tags": [
                    "template-controller"
                ],
                "operationId": "getTemplateFilesByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "byte"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/templates/data/{id}": {
            "get": {
                "tags": [
                    "template-controller"
                ],
                "operationId": "getTemplateById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemplateDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/templates/byType/{type}": {
            "get": {
                "tags": [
                    "template-controller"
                ],
                "operationId": "getFilesByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/templates/delete/{id}": {
            "delete": {
                "tags": [
                    "template-controller"
                ],
                "operationId": "deleteTemplate1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TemplateDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "templateName": {
                        "type": "string"
                    },
                    "templateType": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "templateFile": {
                        "type": "string",
                        "format": "binary"
                    },
                    "templateFileName": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "string"
                    },
                    "productId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updateDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "RefreshAuthDto": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
            "DepartmentInfo": {
                "type": "object",
                "properties": {
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "departmentName": {
                        "type": "string"
                    }
                }
            },
            "DesignationBasedOperation": {
                "type": "object",
                "properties": {
                    "desginationId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "designationName": {
                        "type": "string"
                    },
                    "operationMapped": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OperationMapped"
                        }
                    }
                }
            },
            "LocationInfo": {
                "type": "object",
                "properties": {
                    "locationId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "locationName": {
                        "type": "string"
                    }
                }
            },
            "OperationMapped": {
                "type": "object",
                "properties": {
                    "opeartionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "operationType": {
                        "type": "string"
                    },
                    "operationName": {
                        "type": "string"
                    }
                }
            },
            "ProductCategory": {
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "productName": {
                        "type": "string"
                    }
                }
            },
            "UserInfo": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "locationInfo": {
                        "$ref": "#/components/schemas/LocationInfo"
                    },
                    "departmentInfo": {
                        "$ref": "#/components/schemas/DepartmentInfo"
                    },
                    "productCategory": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductCategory"
                        }
                    },
                    "designationBasedOperation": {
                        "$ref": "#/components/schemas/DesignationBasedOperation"
                    }
                }
            },
...
            "AuthenticationDto": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}